+++ cli/csstools-cli/src/plugins/postcss-double-position-gradients.ts
@@ -4,15 +4,16 @@ import { cli, helpTextLogger } from '@csstools/base-cli';
 export default function postcssDoublePositionGradients() {
 	cli(
 		plugin,
-		['preserve'],
+		['preserve', 'enableProgressiveCustomProperties'],
 		helpTextLogger(
 			'@csstools/cli postcss-double-position-gradients',
 			'PostCSS Double Position Gradients',
-			'Lets you use double-position gradients in CSS, following the [CSS Image Values and Replaced Content] specification',
+			'Lets you use double-position gradients in CSS, following the CSS Image Values and Replaced Content specification',
 			{
 				dir: 'ltr',
 				preserve: true,
 				shadow: true,
+				enableProgressiveCustomProperties: false,
 			},
 		),
 		false,