+++ cli/csstools-cli/src/plugins/postcss-oklab-function.ts
@@ -1,14 +1,14 @@
-import plugin from 'postcss-lab-function';
+import plugin from 'postcss-color-functional-notation';
 import { cli, helpTextLogger } from '@csstools/base-cli';
 
-export default function postcssLabFunction() {
+export default function postcssOKLabFunction() {
 	cli(
 		plugin,
 		['preserve', 'enableProgressiveCustomProperties', 'subFeatures'],
 		helpTextLogger(
-			'@csstools/cli postcss-lab-function',
-			'PostCSS Lab function',
-			'Convert lab() to rgb()',
+			'@csstools/cli postcss-oklab-function',
+			'PostCSS OKLab Function',
+			'Lets you use oklab() and oklch() color functions in CSS.',
 			{
 				preserve: true,
 				enableProgressiveCustomProperties: false,